requirements.txt
Django==3.2.7
calculations==1.10
dataclasses==0.8
djangorestframework==3.12.4
utils==1.0.1
